feat: 新增教材预订

+++ src/views/system/textbookReservation/index.vue
@@ -5,7 +5,6 @@
       :biz-buttons="bizButtons"
       :common-buttons="commonButtons"
       :data.sync="listQuery"
-      :extra-dicts="extraDicts"
       @handle-button-event="handleButtonEvent"
     />
     <div style="padding: 20px">
@@ -17,22 +16,26 @@
         :refresh="tableInfo.refresh"
         :field-list="tableInfo.fieldList"
         :data.sync="tableInfo.data"
-        export-file-name="教材订单表"
+        :table-index="true"
+        :single-select="true"
+        export-file-name="教材预订表"
         @selection-change="selectionChange"
-      >
-        <template #col-handle="{ row }">
-          <sv-button type="text" @click="handelDelete(row)">删除</sv-button>
-        </template>
-      </sv-table>
+      />
     </div>
+    <sv-excel-import
+      ref="svExcelImport"
+      :visible.sync="dialogExcelVisible"
+      @excel-import="excelImport"
+    />
   </div>
 </template>
 
 <script>
-import { delObj, pageQuery } from './api'
+import { batch, delObj, pageQuery } from './api'
+import { parseTime } from '@/utils'
 
 export default {
-  name: 'TextbookOrder',
+  name: 'TextbookReservation',
   data() {
     return {
       pageQuery,
@@ -44,13 +47,15 @@ export default {
           label: '订单编号'
         },
         {
-          type: 'input',
+          type: 'select',
           prop: 'year',
+          dictType: 'year',
           label: '年级'
         },
         {
-          type: 'input',
+          type: 'select',
           prop: 'major',
+          dictType: 'major',
           label: '专业'
         },
         {
@@ -62,26 +67,6 @@ export default {
           type: 'input',
           prop: 'textbookName',
           label: '教材名字'
-        },
-        {
-          type: 'input',
-          prop: 'count',
-          label: '教材数量'
-        },
-        {
-          type: 'input',
-          prop: 'price',
-          label: '教材单价'
-        },
-        {
-          type: 'input',
-          prop: 'orderStatus',
-          label: '订单状态'
-        },
-        {
-          type: 'input',
-          prop: 'crtTime',
-          label: '创建时间'
         }
       ],
       // 搜索条件
@@ -91,18 +76,37 @@ export default {
       },
       // 业务按钮
       bizButtons: [
-        //
+        {
+          name: '新增',
+          show: true,
+          event: 'add'
+        },
+        {
+          name: '编辑',
+          show: true,
+          event: 'edit'
+        },
+        {
+          name: '删除',
+          show: true,
+          event: 'delete'
+        }
       ],
       // 通用按钮
       commonButtons: [
         {
-          name: '导出全部',
-          event: 'exportAll',
+          name: '导入模板下载',
+          event: 'importDownload',
           show: true
         },
         {
-          name: '导出当前页',
-          event: 'exportCurrentPage',
+          name: '导入',
+          event: 'import',
+          show: true
+        },
+        {
+          name: '导出全部',
+          event: 'exportAll',
           show: true
         }
       ],
@@ -115,61 +119,57 @@ export default {
           {
             label: '订单编号',
             prop: 'orderCode',
-            minWidth: 150
+            minWidth: 120
           },
           {
             label: '年级',
             prop: 'year',
-            minWidth: 150
+            dictType: 'year',
+            minWidth: 80
           },
           {
             label: '专业',
             prop: 'major',
-            minWidth: 150
+            dictType: 'major',
+            minWidth: 120
           },
           {
             label: '教材编号',
             prop: 'textbookCode',
-            minWidth: 150
+            minWidth: 120
           },
           {
             label: '教材名字',
             prop: 'textbookName',
-            minWidth: 150
-          },
-          {
-            label: '教材数量',
-            prop: 'count',
-            minWidth: 150
+            minWidth: 120
           },
           {
             label: '教材单价',
             prop: 'price',
-            minWidth: 150
+            minWidth: 60
           },
           {
+            label: '教材数量',
+            prop: 'count',
+            minWidth: 60
+          },
+          {
+            type: 'tag',
             label: '订单状态',
             prop: 'orderStatus',
-            minWidth: 150
+            dictType: 'orderStatus',
+            minWidth: 80
           },
           {
             label: '创建时间',
-            prop: 'crtTime',
-            minWidth: 150
-          },
-          {
-            type: 'slot',
-            label: '操作',
-            prop: 'handle',
-            minWidth: 150
+            prop: 'crtTime'
           }
         ],
         // 数据
         data: []
       },
-      // 手动字典
-      extraDicts: {},
-      currentRow: {}
+      currentRow: {},
+      dialogExcelVisible: false
     }
   },
   computed: {
@@ -187,7 +187,53 @@ export default {
     selectionChange(row) {
       this.currentRow = row
     },
-    async handelDelete(data) {
+    handleButtonEvent(event) {
+      switch (event) {
+        case 'search':
+          this.getList()
+          break
+        case 'add':
+          this.add()
+          break
+        case 'edit':
+          this.edit()
+          break
+        case 'delete':
+          this.delete()
+          break
+        case 'importDownload':
+          window.open('/static/importTemp/预订批量导入模板.xlsx', '_blank')
+          break
+        case 'import': // 导入
+          this.dialogExcelVisible = true
+          break
+        case 'exportAll': // 导出全部
+          this.$refs.svTable.exportExcel()
+          break
+        default:
+          break
+      }
+    },
+    add() {
+      this.$openTag(this, {
+        name: 'TextbookReservationDetail',
+        title: '教材预订明细',
+        id: null,
+        status: 'add'
+      })
+    },
+    edit() {
+      if (!this.$refs.svTable.checkRow()) return
+
+      this.$openTag(this, {
+        name: 'textbookReservationDetail',
+        title: '教材预订明细',
+        id: this.currentRow[0].id,
+        status: 'edit'
+      })
+    },
+    async delete() {
+      if (!this.$refs.svTable.checkRow()) return
       const message = '删除后将不可再恢复，谨慎操作，请确认是否需要删除？'
       await this.$confirm(message, '删除', {
         type: 'delete',
@@ -195,28 +241,44 @@ export default {
         cancelButtonText: '返回'
       })
         .then(async() => {
-          await delObj(data.id)
-          this.svMessage({
-            type: 'success',
-            title: '删除成功'
-          })
+          await delObj(this.currentRow.id)
+
           this.getList()
+          this.$notify({
+            title: '成功',
+            type: 'success',
+            message: '操作成功',
+            duration: 2000
+          })
         })
     },
-    handleButtonEvent(event) {
-      switch (event) {
-        case 'search':
+    async excelImport(data) {
+      const list = []
+      data.forEach(item => {
+        const excelData = {}
+        excelData.year = item['年级']
+        excelData.major = item['专业']
+        excelData.textbookCode = item['教材编号']
+        excelData.count = item['教材数量']
+        excelData.crtTime = parseTime(new Date())
+        list.push(excelData)
+      })
+
+      const res = await batch(list)
+      try {
+        if (res.status === 200) {
           this.getList()
-          break
-        case 'exportAll': // 导出全部
-          this.$refs.svTable.exportExcel()
-          break
-        case 'exportCurrentPage': // 导出当前页
-          this.$refs.svTable.exportCurrentPage()
-          break
-        default:
-          break
+          this.$notify({
+            title: '成功',
+            type: 'success',
+            message: '添加成功',
+            duration: 2000
+          })
+        }
+      } catch (e) {
+        this.dialogExcelVisible = false
       }
+      this.dialogExcelVisible = false
     }
   }
 }
